pyproject.toml
[project]
name = "flow-py-connect"
version = "0.1.0"
description = "Flow Connect - GraphQL FastAPI Backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "strawberry-graphql[debug-server, fastapi]>=0.278.0",
    "pydantic-settings>=2.2.1,<3",
    "fastapi[standard]>=0.110.0",
    "sqlalchemy>=2.0.30,<3",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.10,<3",
    "alembic>=1.13.1,<2",
    "pydantic>=2.7.1,<3",
    "aioinject==1.10.2",
    "orjson>=3.10.5,<4",
    "pendulum>=3.0.0,<4",
    "loguru>=0.7.2,<0.8",
    "pyjwt[crypto]>=2.10.1",
    "python-dotenv>=1.0.0",
    "greenlet>=3.2.4",
    "cryptography>=46.0.2",
    "uvicorn[standard]>=0.35.0",
    "flow-py-commons",
    "httpx>=0.27.0",
]

[dependency-groups]
dev = [
    "bandit[toml]>=1.7.8,<2",
    "isort>=7.0.0",
    "ruff>=0.6.3",
    "mypy-extensions>=1.0.0,<2",
    "mypy>=1.10.0,<2",
    "basedpyright>=1.12.2,<2",
    "deptry>=0.24.0",
    "pytest>=9.0.2",
    "httpx>=0.27.0",
    "pytest-asyncio>=0.25.3",
    "pre-commit>=4.3.0",
    "faker>=33.1.0",
    "testcontainers>=4.13.1",
]

[tool.uv.sources]
flow-py-commons = { git = "ssh://git@github.com/FlowRMS/flowbot-commons.git", tag = "1.23.00" }

[tool.pytest.ini_options]
asyncio_mode = "auto"
